+++ ui/src/features/annotator/annotations/editable-annotation.component.tsx
@@ -5,7 +5,7 @@ import { ReactNode } from 'react';
 
 import { useZoom } from '../../../components/zoom/zoom';
 import { useSelectedAnnotations } from '../select-annotation-provider.component';
-import { EditBoundingBox } from '../tools/bounding-box-tool/bounding-box-tool.component';
+import { EditBoundingBox } from '../tools/edit-bounding-box/edit-bounding-box.component';
 import { Annotation, Rect } from '../types';
 import { useAnnotation } from './annotation.component';
