+++ ui/src/features/annotator/annotator-provider.component.tsx
@@ -39,7 +39,7 @@ export const AnnotatorProvider = ({ mediaItem, children }: { mediaItem: MediaIte
             {
                 shape,
                 id: uuid(),
-                labels: [{ id: uuid(), name: 'Default label', color: '--annotation-fill', isPrediction: false }],
+                labels: [{ id: uuid(), name: 'Default label', color: 'var(--annotation-fill)', isPrediction: false }],
             },
         ]);
     };
