--- ui/src/features/annotator/tools/annotator-drawing-box.component.tsx
@@ -1,20 +1,0 @@
-// Copyright (C) 2025 Intel Corporation
-// SPDX-License-Identifier: Apache-2.0
-
-import { useZoom } from '../../../components/zoom/zoom';
-import { useAnnotator } from '../annotator-provider.component';
-import { DrawingBox } from './drawing-box-tool/drawing-box.component';
-
-export const AnnotatorDrawingBox = () => {
-    const { mediaItem, addAnnotation } = useAnnotator();
-    const { scale: zoom } = useZoom();
-
-    return (
-        <DrawingBox
-            roi={{ x: 0, y: 0, width: mediaItem.width, height: mediaItem.height }}
-            image={new ImageData(mediaItem.width, mediaItem.height)}
-            zoom={zoom}
-            onComplete={addAnnotation}
-        />
-    );
-};
