+++ ui/src/features/annotator/tools/edit-bounding-box/anchor.component.tsx
