+++ ui/src/features/annotator/tools/edit-bounding-box/resize-anchor.component.tsx
