+++ ui/src/features/annotator/tools/edit-bounding-box/translate-shape.component.tsx
