+++ ui/src/features/annotator/tools/utils.ts
@@ -1,7 +1,17 @@
 // Copyright (C) 2025 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-import { Point, RegionOfInterest } from '../../types';
+import { Point, RegionOfInterest } from '../types';
+
+type ElementType = SVGElement | HTMLDivElement;
+export const getRelativePoint = (element: ElementType, point: Point, zoom: number): Point => {
+    const rect = element.getBoundingClientRect();
+
+    return {
+        x: Math.round((point.x - rect.left) / zoom),
+        y: Math.round((point.y - rect.top) / zoom),
+    };
+};
 
 export const getBoundingBoxInRoi = (boundingBox: RegionOfInterest, roi: RegionOfInterest) => {
     const x = Math.max(0, boundingBox.x);
